+++ lib/dradis/plugins.rb
@@ -11,10 +11,19 @@ module Dradis
         @@extensions
       end
 
+      # Returns an array of modules representing currently enabled engines
+      def enabled_list
+        @@enabled_list ||= @@extensions.select(&:enabled?)
+      end
+
+      def clear_enabled_list
+        @@enabled_list = nil
+      end
+
       # Filters the list of plugins and only returns those that provide the
-      # requested feature.
+      # requested feature and enabled
       def with_feature(feature)
-        @@extensions.select do |plugin|
+        enabled_list.select do |plugin|
           # engine = "#{plugin}::Engine".constantize
           plugin.provides?(feature)
         end